Move DB functions into own file
